+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ---------------
 
+2.11.1 (2025-12-17)
++++++++++++++++++++
+
+**Bug Fixes**
+
+- 3.14 utc market fix
+- seconds_to_start bug when set to 0 seconds
+- limit order persistence validation (MarketOnClose when market is inplay or no inplay available)
+- correctly update MarketOnClose/LimitOnClose as EXECUTION_COMPLETE once processed
+
 2.11.0 (2025-12-11)
 +++++++++++++++++++